+++ src/hooks/useCreateMessagesUi.tsx
@@ -3,35 +3,7 @@ import { useGetMessages } from './useGetMessages';
 import { GoogleMessage } from '../models/GoogleMessage';
 import { MessageItem } from '../components/MessageItem/MessageItem';
 import { UserContext } from '../App';
-
-export const sortMessageData = (messageData: GoogleMessage) => {
-  const result = {
-    from: '',
-    snippet: messageData.snippet,
-    title: '',
-    date: '',
-    text: messageData.payload.parts[0].body.data,
-  };
-  messageData.payload.headers.forEach(
-    (element: { name: string; value: string }) => {
-      switch (element.name) {
-        case 'From':
-          result.from = element.value;
-          break;
-        case 'Subject':
-          result.title = element.value;
-          break;
-        case 'Date':
-          const messageDate = new Date(element.value);
-          result.date = `${messageDate.getDate()}-${messageDate.getMonth()}-${messageDate.getFullYear()}`;
-          break;
-        default:
-          break;
-      }
-    },
-  );
-  return result;
-};
+import { sortMessageData } from './useSortMessageData';
 
 export const useCreateMessagesUi = () => {
   const userData = useContext(UserContext);
